requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
dataclasses-json==0.6.7
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filetype==1.2.0
frozenlist==1.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-auth==2.40.3
google-genai==1.33.0
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-google-genai==2.1.10
langchain-text-splitters==0.3.11
langsmith==0.4.25
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
numpy==2.3.2
orjson==3.11.3
packaging==25.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
tenacity==9.1.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
zstandard==0.24.0
